feat: binding logic in observation beta branch on add product item feature
